MOAC, mother of all commits

+++ src/server/routes/contacts.js
@@ -3,8 +3,20 @@ const {renderError} = require('../utils')
 
 const router = require('express').Router()
 
+// router.use( (request, response, next) =>{
+//   console.log(request.session);
+//   if(!request.session) {
+//     response.redirect('/login')
+//
+//   }
+//   next()
+// })
 router.get('/new', (request, response) => {
-  response.render('new')
+  if (request.session.role !== 'admin') {
+    response.sendStatus(403)
+  } else {
+    response.render('new')
+  }
 })
 
 router.post('/', (request, response, next) => {
@@ -17,11 +29,12 @@ router.post('/', (request, response, next) => {
 })
 
 router.get('/:contactId', (request, response, next) => {
+  let userinfo = request.session
   const contactId = request.params.contactId
   if (!contactId || !/^\d+$/.test(contactId)) return next()
   DbContacts.getContact(contactId)
     .then(function(contact) {
-      if (contact) return response.render('show', { contact })
+      if (contact) return response.render('show', { contact, userinfo })
       next()
     })
     .catch( error => renderError(error, response, response) )
@@ -29,13 +42,17 @@ router.get('/:contactId', (request, response, next) => {
 
 
 router.get('/:contactId/delete', (request, response, next) => {
-  const contactId = request.params.contactId
-  DbContacts.deleteContact(contactId)
-    .then(function(contact) {
-      if (contact) return response.redirect('/')
-      next()
-    })
-    .catch( error => renderError(error, response, response) )
+  if (request.session.role !== 'admin') {
+    response.sendStatus(403);
+  } else {
+    const contactId = request.params.contactId
+    DbContacts.deleteContact(contactId)
+      .then(function(contact) {
+        if (contact) return response.redirect('/')
+        next()
+      })
+      .catch( error => renderError(error, response, response) )
+  }
 })
 
 router.get('/search', (request, response, next) => {
